Extension slot added for visit summary
