+++ packages/esm-active-visits-app/src/visits-summary/visits-components/encounter-observations.component.tsx
