+++ packages/esm-active-visits-app/src/visits-summary/visits-components/medications-summary.component.tsx
