+++ packages/esm-active-visits-app/src/visits-summary/visits-components/notes-summary.component.tsx
