+++ packages/esm-active-visits-app/src/visits-summary/visits-components/tests-summary.component.tsx
@@ -1,22 +1,16 @@
 import React from 'react';
 import { ExtensionSlot } from '@openmrs/esm-framework';
-import { ExternalOverviewProps } from '@openmrs/esm-patient-common-lib';
 import { Encounter } from '../visit.resource';
 
 const TestsSummary = ({ patientUuid, encounters }: { patientUuid: string; encounters: Array<Encounter> }) => {
-  const filter = React.useMemo<ExternalOverviewProps['filter']>(() => {
+  const filter = React.useMemo(() => {
     const encounterIds = encounters.map((e) => `Encounter/${e.uuid}`);
     return ([entry]) => {
       return encounterIds.includes(entry.encounter.reference);
     };
   }, [encounters]);
 
-  return (
-    <ExtensionSlot
-      extensionSlotName="test-results-filtered-overview"
-      state={{ filter, patientUuid } as ExternalOverviewProps}
-    />
-  );
+  return <ExtensionSlot extensionSlotName="test-results-filtered-overview" state={{ filter, patientUuid }} />;
 };
 
 export default TestsSummary;
